fixed foundation to prevent some file io exceptions

+++ foundation.py
@@ -88,19 +88,23 @@ if __name__ == '__main__':
     logger.info(f"defined ignore file list [{ignore_files}] for files not to copy to destination")
     for repo_resource in dest_repo.tree().traverse():
         resource_name = repo_resource.name
-        dest_file_name = args.dest + resource_name
-        logger.info(f"naming {dest_file_name} as the destination file and path for {resource_name}")
-        if os.path.isdir(resource_name):
-            logger.info(f"found a folder in source resources {resource_name}, copying as a folder {dest_file_name}")
-            shutil.copytree(resource_name, dest_file_name)
-        elif resource_name == 'readme_template.md':
-            logger.info(f"converting readme template to readme: {resource_name} to README.md")
-            safe_copy('readme_template.md', args.dest + 'README.md')
-        elif resource_name in ignore_files:
-            logger.info(f"ignoring {resource_name} from copy to destination")
-        else:
-            safe_copy(resource_name, dest_file_name)
-            if resource_name == "AUTHORS.md":
-                append_authors(dest_repo, args.dest)
-            if resource_name == "LICENSE":
-                update_license(dest_repo, args.dest)
+        if os.path.exists(repo_resource.name):
+            dest_file_name = args.dest + resource_name
+            logger.info(f"naming {dest_file_name} as the destination file and path for {resource_name}")
+            if os.path.isdir(resource_name):
+                logger.info(f"found a folder in source resources {resource_name}, copying as a folder {dest_file_name}")
+                if not os.path.exists(dest_file_name):
+                    shutil.copytree(resource_name, dest_file_name)
+                else:
+                    logger.info(f"Not copying folder {dest_file_name} because it already exists")
+            elif resource_name == 'readme_template.md':
+                logger.info(f"converting readme template to readme: {resource_name} to README.md")
+                safe_copy('readme_template.md', args.dest + 'README.md')
+            elif resource_name in ignore_files:
+                logger.info(f"ignoring {resource_name} from copy to destination")
+            else:
+                safe_copy(resource_name, dest_file_name)
+                if resource_name == "AUTHORS.md":
+                    append_authors(dest_repo, args.dest)
+                if resource_name == "LICENSE":
+                    update_license(dest_repo, args.dest)
